Update README.md
i just took out the note-taking and insurance filing support under the 'options include' in this ReadMe file bc I think for the MVP we emphasize therapist-facing tools - this mitigates additional risk management with client facing admin needs but i'm open to us discussing a tool that toggles from CLIENT facing to THERAPIST facing and have thoughts on that including (if a client opts-in) a range of tools the therapist could co-create with the client: relational strength training bot that is tailored to client's attachment style, a curated conversation ally based on client strengths and preferences (?), Ai generated clinically proven interventions for "HW" outside of session across the range of modalities (this could streamline theoretically directly into the client's 'chart' if this was interfacing with an EHR for example...) but for a first copilot iteration it seems unique and good to center the provider IMHO - and again I'm flexible! open to market preference!